Move `killAll()` to linux specific file
